+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+v4.0.0
+===
+
+Breaking
+---
+ - Now it uses `render` method in props instead of `children`
+
+Bug fixes
+---
+ - Fixed double subscribing to stream
+
+New Features
+---
+ - Use `debounce` prop to debounce updating component for new values in streams. Useful for atomic updates of multiple streams bound to one component
+
 v3.0.0
 ===